requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
environ==1.0
idna==2.10
Markdown==3.3.4
mysqlclient==2.0.3
phonenumbers==8.12.20
Pillow==8.3.1
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
